Modifications #1705:
- my_pdt_fromstring() : This function has been renamed to pdt_fromisoformat() such to allign to what it replaces (.fromisoformat()).

- _utils.py : The newly named pdt_fromisoformat() function has been moved from constraints.py into utils.py, such that it can be used in conjunction to multiple different operators.

- test_utils.py : The tests made for the pdt_fromisoformat() function have been moved to test_utils.py, to reflect the relocation of the function from constraints.py to _utils.py.

- test_read.py : This file has been modified to handle the changes to name and location of the pdt_fromisoformat() function.
